+++ test/Candoumbe.Types.Strings.UnitTests/ReplaceStringWithString.cs
@@ -3,33 +3,35 @@ using FsCheck;
 using FsCheck.Experimental;
 using FsCheck.Fluent;
 
-namespace Candoumbe.Types.UnitTests.Strings;
+namespace Candoumbe.Types.Strings.UnitTests.Strings;
 
 /// <summary>
-/// Operation of replacing a <see langword="char"/> by another.
+/// Operation of replacing a <see langword="string"/> by another.
 /// </summary>
-internal class ReplaceCharWithChar : Operation<StringSegmentLinkedList, StringSegmentLinkedListState>
+internal class ReplaceStringWithString : Operation<StringSegmentLinkedList, StringSegmentLinkedListState>
 {
-    private readonly char _oldValue;
-    private readonly char _newValue;
+    private readonly string _oldValue;
+    private readonly string _newValue;
 
     /// <summary>
-    /// Builds a new <see cref="ReplaceCharWithChar"/> instance.
+    /// Builds a new <see cref="ReplaceStringWithString"/> instance.
     /// </summary>
     /// <param name="oldValue">The character to replace</param>
     /// <param name="newValue">The character that will replace <paramref name="oldValue"/></param>
-    internal ReplaceCharWithChar(char oldValue, char newValue)
+    internal ReplaceStringWithString(string oldValue, string newValue)
     {
         _oldValue = oldValue;
         _newValue = newValue;
     }
 
     /// <inheritdoc />
+    public override bool Pre(StringSegmentLinkedListState currentState) => currentState.Value.Length > 0;
+
+    /// <inheritdoc />
     public override Property Check(StringSegmentLinkedList sut, StringSegmentLinkedListState state)
     {
-        sut = sut.Replace(_oldValue, _newValue);
-        string actual = sut.ToStringValue();
-
+        StringSegmentLinkedList result  = sut.Replace(_oldValue, _newValue);
+        string actual = result.ToStringValue();
         return ( actual == state.Value ).Label($"""
                                                 Actual : "{actual}"
                                                 Expect : "{state.Value}"
@@ -38,8 +40,8 @@ internal class ReplaceCharWithChar : Operation<StringSegmentLinkedList, StringSe
 
     /// <inheritdoc />
     public override StringSegmentLinkedListState Run(StringSegmentLinkedListState state)
-        => state with {Value = state.Value.Replace(_oldValue, _newValue)};
+        => state with { Value = state.Value.Replace(_oldValue, _newValue) };
 
     /// <inheritdoc />
-    public override string ToString() => $"replace char '{_oldValue}' (ascii: {(int)_oldValue}) with char '{_newValue}' (ascii: {(int)_newValue})";
+    public override string ToString() => $@"replace string ""{_oldValue}"" with string ""{_newValue}""";
 }